♻️(backend) move abilities computation to the permissions backend
Pure move of Item.get_abilities into the role backend so every
permission decision sits behind the same facade.
